# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: TRACKSY API
  description: |
    Smart Student Transport Tracking System API Contract
    
    This API provides endpoints for:
    - Student mobile application
    - Driver mobile application
    - Admin dashboard
    
    All endpoints require authentication except login and registration.
  version: 1.0.0
  contact:
    name: TRACKSY Team
    email: support@tracksy.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: http://127.0.0.1:4010
    description: Prism mock server
  - url: https://api.tracksy.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        busId:
          type: integer
        seatNumber:
          type: string
        tripDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        createdAt:
          type: string
          format: date-time
    BookingCreate:
      type: object
      required:
        - busId
        - seatNumber
        - tripDate
      properties:
        busId:
          type: integer
        seatNumber:
          type: string
        tripDate:
          type: string
          format: date
...
paths:
...
  /bookings:
    get:
      tags:
        - Bookings
        - Students
      summary: Get user bookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                  pagination:
                    type: object
    post:
      tags:
        - Bookings
        - Students
      summary: Create booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Seat not available or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...